requirements.txt
Flask
pypdf
google-generativeai
openpyxl
streamlit
pandas
